Implement and refactor the framework for pg function supporting (#59)
* refactor regobject and define regproc

* define pg_proc table

* handle function name with argument and return type pattern

* fix regobject listing

* implement pg functions

* fix table function rewrite for bigquery

* register pg functions in calcite manually

* address reviews

+++ core/cml-postgres-wire-protocol/src/main/java/io/cml/sql/RegObjectInterpreter.java
@@ -63,7 +63,7 @@ public class RegObjectInterpreter
                 if (showObjectAsName) {
                     return new StringLiteral(regObject.getName());
                 }
-                return new LongLiteral(Integer.toString(regObject.getOid()));
+                return new LongLiteral(Long.toString(regObject.getOid()));
             }
             return node;
         }
@@ -82,7 +82,7 @@ public class RegObjectInterpreter
             if (showObjectAsName) {
                 return new StringLiteral(regObject.getName());
             }
-            return new LongLiteral(Integer.toString(regObject.getOid()));
+            return new LongLiteral(Long.toString(regObject.getOid()));
         }
 
         @Override
